% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conservatism.R
\name{Conservatism}
\alias{Conservatism}
\title{placeholder}
\usage{
Conservatism(network_list, n_it, species, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
placeholder
}
\examples{
add(1, 1)
}
