% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RolePredict.R
\name{RolePredict}
\alias{RolePredict}
\title{placeholder}
\usage{
RolePredict(
  training_networks,
  networks_to_predict,
  conservatism,
  weighting = NULL,
  n_it,
  species_remapping = NULL,
  ...
)
}
\arguments{
\item{training_networks}{A number.}

\item{networks_to_predict}{A number.}

\item{conservatism}{sss}

\item{weighting}{ddd}

\item{n_it}{ssss}

\item{species_remapping}{ddd}

\item{...}{ddd}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
placeholder
}
\examples{
#
}
