% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateRolesMultipleNetworks.R
\name{CalculateRolesMultipleNetworks}
\alias{CalculateRolesMultipleNetworks}
\title{placeholder}
\usage{
CalculateRolesMultipleNetworks(network_list, ...)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
placeholder
}
\examples{
add(1, 1)
}
