% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConservatismPermTest.R
\name{ConservatismPermTest}
\alias{ConservatismPermTest}
\title{placeholder}
\usage{
ConservatismPermTest(roles, n_it, species)
}
\arguments{
\item{roles}{A number.}

\item{n_it}{A number.}

\item{species}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
placeholder
}
\examples{
#
}
